fyp24-container/requirements.txt
### FontDiffuser (original) ###
transformers>=4.33.1,<4.34.0
accelerate>=0.23.0,<0.24.0
diffusers>=0.23.0,<0.24.0
huggingface_hub>=0.25.0,<0.26.0
pyyaml
pygame
opencv-python
info-nce-pytorch
kornia>=0.7.0,<0.7.4  # Do not use 0.7.4 or 0.8.0; other versions are fine

### FYP24 Team ###

## Training ##

# for tensorboard
tensorflow>=2.17.0,<2.18.0

## Evaluation ##
lpips

# for torcheval
# https://github.com/pytorch/torcheval/blob/main/dev-requirements.txt
torcheval-nightly==2024.7.31
scikit-image>=0.18.0,<0.19.0

### Generative Calligraphy Website ###
fastapi[standard]>=0.115.8,<0.116.0
pytest>=8.4.1,<9.0.0
pytest-asyncio>=1.1.0,<2.0.0
pytest-timeout>=2.4.0,<3.0.0
